test(landing): move the #2996 review-verdict reproductions into tests/regression
Extract the three #2996 red-first P0 reproductions out of files that
otherwise carry unrelated, passing tests, so those files stop colliding
with non-regression CI selections that were previously red only because
they collected an intentionally-failing pin:

- tests/integration/test_review_cycle_rejection_only.py::test_approving_a_rejected_wp_writes_no_verdict_artifact
- tests/review/test_cycle.py::test_self_referential_feedback_source_is_rejected
- tests/review/test_cycle.py::test_new_cycle_body_never_duplicates_a_prior_cycle_file

into tests/regression/test_issue_2996_approval_after_rejection_writes_no_verdict.py
and tests/regression/test_issue_2996_self_referential_and_duplicate_feedback.py,
each carrying its own copy of the fixtures/helpers it needs. Root-cause
docstrings and assertions are preserved verbatim; the three tests still
reproduce their #2996 defect (RED), and the source files' remaining tests
still pass.
